requirements.txt
langchain==0.0.234
openai==0.27.8
python-dotenv==1.0.0
streamlit==1.24.1
llama-cpp-python==0.1.65
transformers==4.32.1
PyPDF2==3.0.1
tiktoken==0.4.0
qdrant-client==1.3.1